refactor: add showItemInClient helper
