+++ packages/api-reference/src/helpers/index.ts
@@ -15,12 +15,13 @@ export * from './getVariableNames'
 export * from './hasModels'
 export * from './hasSecuritySchemes'
 export * from './isJsonString'
+export * from './isValidUrl'
 export * from './json2xml'
 export * from './mapFromObject'
 export * from './objectMerge'
 export * from './prettyPrintJson'
 export * from './replaceVariables'
 export * from './scrollToId'
+export * from './showItemInClient'
 export * from './splitMarkdownInSections'
 export * from './transformToJson'
-export * from './isValidUrl'
